+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 5.7.0.0 - 2025-06-XX
 * Added ability to move a target from one project to another.
+* Added display of the thumbnail image to the acquired image detail view.
+* Added ability to manually change the grading status on acquired images.
 
 ## 5.6.5.0 - 2025-06-20
 * Fixed issue where TS Sync Condition was not resetting properly.